+++ app/projects/[id]/welcome/layout.js
@@ -21,12 +21,20 @@ export default async function WelcomeLayout({ params, children }) {
 
   return (
     <Providers project={project} session={session}>
-      <SiteHeader hideLogo />
+      <SiteHeader hideLogo>
+        <Link
+          href={`/projects/${project.id}`}
+          target="_blank"
+          className="hover:underline"
+        >
+          Manage Project
+        </Link>
+      </SiteHeader>
       <div className="flex-col items-center pt-16 space-y-4 sm:flex-row sm:px-6">
         <div className="text-3xl font-bold text-center">
           <Link
             href={`/projects/${project.id}/welcome`}
-            className="hover:text-tertiary"
+            className="hover:underline"
           >
             {project.name}
           </Link>
