+++ src/components/ui/SiteHeader.js
@@ -1,7 +1,6 @@
 "use client";
 
 import { React, useState } from "react";
-import { createPortal } from "react-dom";
 import { useSession } from "next-auth/react";
 import Link from "next/link";
 import Image from "next/image";
@@ -9,18 +8,12 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { signOut } from "next-auth/react";
 import classnames from "classnames";
 
-import Modal from "src/components/ui/Modal";
-import ThemeSelector from "src/components/ui/ThemeSelector";
 import logo from "public/logo.png";
+import ThemeAction from "src/components/domains/bookmarks/ThemeAction";
 
-export default function SiteHeader({ hideLogo }) {
+export default function SiteHeader({ hideLogo, children }) {
   const { data: session } = useSession();
   const user = session?.user;
-  const [editingTheme, setEditingTheme] = useState(false);
-
-  const toggleEditingTheme = () => {
-    setEditingTheme(!editingTheme);
-  };
 
   return (
     <>
@@ -40,20 +33,14 @@ export default function SiteHeader({ hideLogo }) {
           </div>
         </Link>
         <div className="flex-1" />
-        {editingTheme &&
-          createPortal(
-            <Modal title="Change Theme" close={toggleEditingTheme}>
-              <ThemeSelector />
-            </Modal>,
-            document.body
-          )}
-        <div className="cursor-pointer" onClick={toggleEditingTheme}>
+        <ThemeAction>
           <FontAwesomeIcon icon={["fas", "brush"]} />
-        </div>
+        </ThemeAction>
         {user && !user.fake && (
           <div className="flex flex-col items-end ml-4 space-x-3">
-            <span>{session.user.email}</span>
-            <button className="underline" onClick={signOut}>
+            <div>{session.user.email}</div>
+            {children}
+            <button className="hover:underline" onClick={signOut}>
               Sign out
             </button>
           </div>
